refactor: 알람 관련 UI 이벤트 및 상태 파일 경로 변경
