+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/edit/alarm/AlarmUiEvent.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.view.edit.alarm.listener
+package com.bottari.presentation.view.edit.alarm
 
 sealed interface AlarmUiEvent {
     data object AlarmCreateSuccess : AlarmUiEvent
