+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/edit/alarm/AlarmUiState.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.view.edit.alarm.listener
+package com.bottari.presentation.view.edit.alarm
 
 import com.bottari.presentation.model.AlarmUiModel
